gcloudgrad07/requirements.txt
alembic==1.13.1
blinker==1.8.1
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
confluent-kafka==2.4.0
dnspython==2.6.1
dominate==2.9.1
email_validator==2.1.1
Flask==3.0.3
Flask-Bootstrap==3.3.7.1
Flask-Login==0.6.3
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
greenlet==3.0.3
idna==3.7
install==1.3.5
itsdangerous==2.2.0
Jinja2==3.1.3
kafka-python==2.0.2
Mako==1.3.3
MarkupSafe==2.1.5
paho-mqtt==2.1.0
psycopg2-binary==2.9.9
requests==2.31.0
SQLAlchemy==2.0.29
typing_extensions==4.11.0
urllib3==2.2.1
visitor==0.1.3
waitress==3.0.0
Werkzeug==3.0.2
WTForms==3.1.2
